add env variable
