+++ front/src/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useCallback, useEffect } from 'react';
+import React from 'react';
 import { BrowserRouter as Router, Redirect, Route, Switch } from 'react-router-dom';
 import MainNavigation from './components/Navigation/MainNavigation';
 import "./scss/GlobalStyle.scss";
@@ -11,50 +11,10 @@ import UpdatePlace from './pages/places/UpdatePlace'
 import ConsultaRapida from './pages/consultasRapidas/ConsultaRapida';
 import Auth from './pages/users/Auth';
 import { AuthContext } from './shared/context/auth-context';
-
-let logoutTimer;
+import { useAuth } from './shared/hooks/auth-hook';
 
 const App = () => {
-  const [token, setToken] = useState(false);
-  const [tokenExpirationDates, setTokenExpirationDates] = useState();
-  const [userId, setUserId] = useState();
-
-  const login = useCallback((uid, token, expirationDate) => {
-    setToken(token);
-    setUserId(uid);
-    const tokenExpirationDate = expirationDate || new Date(new Date().getTime() + 1000 * 60 * 60);
-    setTokenExpirationDates(tokenExpirationDate);
-    localStorage.setItem('useData',JSON.stringify(
-      {
-        userId: uid,
-        token: token,
-        expiration: tokenExpirationDate.toISOString()
-      }
-    ))
-  }, []);
-
-  const logout = useCallback(() => {
-    setToken(null);
-    setTokenExpirationDates(null);
-    setUserId(null);
-    localStorage.removeItem('useData')
-  }, []);
-
-  useEffect(()=>{
-    if(token && tokenExpirationDates){
-      const remainingTime = tokenExpirationDates.getTime() - new Date().getTime();
-      logoutTimer = setTimeout(logout, remainingTime)
-    } else {
-      clearTimeout(logoutTimer);
-    }
-  },[token, logout, tokenExpirationDates])
-
-  useEffect(() => {
-    const storedData = JSON.parse(localStorage.getItem('useData'))
-    if(storedData && storedData.token && new Date(storedData.expiration) > new Date()){
-      login(storedData.userId, storedData.token, new Date(storedData.expiration))
-    }
-  },[login])
+  const { token, login, logout, userId } = useAuth();
 
   let routes;
