+++ front/src/components/UsersComponent/PlaceItem.js
@@ -22,12 +22,10 @@ const PlaceItem = (props) => {
         setShowConfirmModal(false);
         try{
             await sendRequest(
-                `http://localhost:5000/api/places/${props.id}`,
+                process.env.REACT_APP_ + `places/${props.id}`,
                 'DELETE',
                 null,
-                {
-                    Authorization: 'Bearer ' + auth.token
-                }
+                { Authorization: 'Bearer ' + auth.token }
             );
             props.onDelete(props.id);
         }catch(err){}
@@ -85,7 +83,7 @@ const PlaceItem = (props) => {
                 <CardMedia
                     component="img"
                     height="140"
-                    image={`http://localhost:5000/${props.image}`}
+                    image={process.env.REACT_APP_IMG + props.image}
                     alt={props.title}
                 />
                 <CardContent>
