+++ front/src/pages/users/Auth.js
@@ -68,7 +68,7 @@ const Auth = () => {
     if(isLoginMode){
         try{
           const responseData = await sendRequest(
-            "http://localhost:5000/api/users/login",
+            process.env.REACT_APP_ + "users/login",
             'POST',
             JSON.stringify({
               email: formState.inputs.email.value,
@@ -93,7 +93,7 @@ const Auth = () => {
         formData.append('image', formState.inputs.image.value)
 
         const responseData = await sendRequest(
-          "http://localhost:5000/api/users/signup",
+          process.env.REACT_APP_ + "users/signup",
           'POST',
           formData
         )
