+++ front/src/pages/users/Users.js
@@ -10,7 +10,7 @@ const Users = () => {
     useEffect(()=>{
         const fetchUsers = async () => {
             try {
-                const response = await sendRequest('http://localhost:5000/api/users');
+                const response = await sendRequest(process.env.REACT_APP_ + 'users');
 
                 setLoadedUsers(response.users);
             } catch(err){}
